requirements.txt
google-adk>=0.3.0
python-dotenv>=1.0
streamlit>=1.39
requests>=2.32.0
